Refactor methods into concern for sharing between presenter and model
